chore(deps): update deps for vue 3 compatibility
This commit updates vue to vue 3 and includes @vue/compat. The compat
package allows for some vue 2 "compatibility" that really just amounts
to vue 3 opt-outs on a app-wide or per-component basis. I've chosen to
not use the opt-outs but rather leverage the compat build's better error
and warning messages to then refactor out codebase. The compat build
should be removed once all errors and warnings are gone.

The commit also introduces some new dependencies that are necessary to
be able to compile our tests. The new vue 3 types don't play nice with
our existing testing packages so this was necessary to fix build-time
errors for the tests.

Finally, the commit updates a number of vue-cli dependencies that still
relied on now deprecated vue packages.