build: create vue config
This vue config file tells webpack, which is what the vue-cli is built
on top of, to alias vue to @vue/compat. This is how we are integrating
the compatibility build into our plugin. This config should be removed
once all warnings and errors from the compat build have been addressed.