feat: use global api to register components
In the vue 3 api, install expects an instance of the createApp return
value. This commit imports the created app and defines it as the type
for the app argument for the install function. Lastly, vue no longer
exports Vue, but rather has you registers components on an existing app.
This commit uses the new global component registration api.