feat: use vue 3 global and composition api in ix-img
This commit refactors the component into a functional component to
remove the vue-class-component dependency. This is necessary because the
dependency has not updated their release candidate for vue 3 support in
months. Furthermore, functional components are quickly becoming the more
idiomatic way to write components in vue 3.

This commit also refactors the component to use the new global
defineComponent api and h composition api. The defineComponent api
exposes a new function, the setup function, that replaces a lot of
the functionality previously done in the render function. Setup can
accepts props as an argument. Setup can also return its own render
function. The new render function for defineComponent does not accept
arguments. And, rather than aliasing createElement to h, vue now
directly exports h as a function which can be used inside the render
function.