test: use vue 3 global api & new testing-library screen api
This commit uses the vue 3 global createApp api instead of Vue.use. It
also uses testing-library's screen API, since the libraries render
function no longer returns a wrapper. The return value of the render
function could alternatively be destructured into {getByTestId} for
example, but this achieves the same thing as using screen. Finally
propsData now gets merged into props by default, so props can just be
used instead.

Note that instead of 'data-testid' we are using 'dataTestId' as a prop.
For reasons I cannot explain the 'data-testid' prop is not getting
passed down to the ix-img component reliably so custom prop had be
defined. This is not desired behavior and will likely need to be
addressed in a future commit.