test: register ix components globally
This commit fixes an issue where `ix-image` `ix-picture` and `ix-source`
were not being registered globally for the test suite. Previously, we
had to work around this by creating a sep. spec for every time we
registred a component. Now, we no longer risk overwriting the `ix-`
imports on a concurrent spec run.