chore: fix rollup error
Rollup kept throwing:  `CompileError: for...of statements are not supported.
Use `transforms: { forOf: false }` to skip transformation and disable this error,
or `transforms: { dangerousForOf: true }` if you know what you're doing
rollup.config.js`

Added the `{ dangerousForOf: true }` to the buble options to fix this.
NOTE: the for...of syntax does not work with IE.